sp_metadata_ir

Struct MetadataIR

pub struct MetadataIR<T: Form = MetaForm> {
    pub pallets: Vec<PalletMetadataIR<T>>,
    pub extrinsic: ExtrinsicMetadataIR<T>,
    pub ty: T::Type,
    pub apis: Vec<RuntimeApiMetadataIR<T>>,
    pub outer_enums: OuterEnumsIR<T>,
}

The intermediate representation for the runtime metadata. Contains the needed context that allows conversion to multiple metadata versions.

Note

Further fields could be added or removed to ensure proper conversion. When the IR does not contain enough information to generate a specific version of the runtime metadata an appropriate default value is used (ie, empty vector).

Fields
pallets: Vec<PalletMetadataIR<T>>

Pallet metadata.

extrinsic: ExtrinsicMetadataIR<T>

Metadata of the extrinsic.

ty: T::Type

The type of the Runtime.

apis: Vec<RuntimeApiMetadataIR<T>>

Metadata of the Runtime API.

outer_enums: OuterEnumsIR<T>

The outer enums types as found in the runtime.