Struct sp_metadata_ir::PalletMetadataIR

pub struct PalletMetadataIR<T: Form = MetaForm> {
    pub name: T::String,
    pub storage: Option<PalletStorageMetadataIR<T>>,
    pub calls: Option<PalletCallMetadataIR<T>>,
    pub event: Option<PalletEventMetadataIR<T>>,
    pub constants: Vec<PalletConstantMetadataIR<T>>,
    pub error: Option<PalletErrorMetadataIR<T>>,
    pub index: u8,
    pub docs: Vec<T::String>,
}

The intermediate representation for a pallet metadata.

Fields
name: T::String

Pallet name.

storage: Option<PalletStorageMetadataIR<T>>

Pallet storage metadata.

calls: Option<PalletCallMetadataIR<T>>

Pallet calls metadata.

event: Option<PalletEventMetadataIR<T>>

Pallet event metadata.

constants: Vec<PalletConstantMetadataIR<T>>

Pallet constants metadata.

error: Option<PalletErrorMetadataIR<T>>

Pallet error metadata.

index: u8

Define the index of the pallet, this index will be used for the encoding of pallet event, call and origin variants.

docs: Vec<T::String>

Pallet documentation.