Struct sp_metadata_ir::SignedExtensionMetadataIR

pub struct SignedExtensionMetadataIR<T: Form = MetaForm> {
    pub identifier: T::String,
    pub ty: T::Type,
    pub additional_signed: T::Type,
}

Metadata of an extrinsic’s signed extension.

Fields
identifier: T::String

The unique signed extension identifier, which may be different from the type name.

ty: T::Type

The type of the signed extension, with the data to be included in the extrinsic.

additional_signed: T::Type

The type of the additional signed data, with the data to be included in the signed payload