Crate sp_metadata_ir

Intermediate representation of the runtime metadata.

Structs

ExtrinsicMetadataIR

Metadata of the extrinsic used by the runtime.

MetadataIR

The intermediate representation for the runtime metadata. Contains the needed context that allows conversion to multiple metadata versions.

OuterEnumsIR

The type of the outer enums.

PalletAssociatedTypeMetadataIR

Metadata of a pallet’s associated type.

PalletCallMetadataIR

Metadata for all calls in a pallet

PalletConstantMetadataIR

Metadata about one pallet constant.

PalletErrorMetadataIR

Metadata about a pallet error.

PalletEventMetadataIR

Metadata about the pallet Event type.

PalletMetadataIR

The intermediate representation for a pallet metadata.

PalletStorageMetadataIR

All metadata of the pallet’s storage. The common prefix used by all storage entries.

PalletViewFunctionMetadataIR

Metadata of a pallet view function method.

PalletViewFunctionParamMetadataIR

Metadata of a pallet view function method argument.

RuntimeApiMetadataIR

Metadata of a runtime trait.

RuntimeApiMethodMetadataIR

Metadata of a runtime method.

RuntimeApiMethodParamMetadataIR

Metadata of a runtime method parameter.

StorageEntryMetadataIR

Metadata about one storage entry.

TransactionExtensionMetadataIR

Metadata of an extrinsic’s signed extension.

Enums

DeprecationInfoIR

Deprecation info for an enums/errors/calls. Denotes full/partial deprecation of the type

DeprecationStatusIR

Deprecation status for an entry inside MetadataIR

StorageEntryModifierIR

A storage entry modifier indicates how a storage entry is returned when fetched and what the value will be if the key is not present. Specifically this refers to the “return type” when fetching a storage entry, and what the value will be if the key is not present.

StorageEntryTypeIR

A type of storage value.

StorageHasherIR

Hasher used by storage maps

Functions

into_latest

Transform the IR to the latest stable metadata version.

into_unstable

Transform the IR to unstable metadata version 16.

into_v14

Transform the IR to metadata version 14.

into_version

Transform the IR to the specified version.

supported_versions

Returns the supported metadata versions.